Move runCommandAsTask to tasks.ts
