+++ extension/terminal.ts
@@ -1,9 +1,13 @@
 
-import * as path from 'path';
 import * as vscode from 'vscode';
 import * as utils from './utilities';
 import { logger } from './logger';
 
+/**
+ * Provides a shared terminal in which you can run commands
+ * 
+ * NOTE: Currently there's no way to wait for a command to complete...
+ */
 export class UefiTerminal implements vscode.Disposable {
     private workspace: vscode.WorkspaceFolder;
     private uefiTerminal : vscode.Terminal = null;
@@ -39,8 +43,6 @@ export class UefiTerminal implements vscode.Disposable {
      * @param args Arguments to pass to the process. 
      * @param cwd  The current working directory for the process
      */
-    // 
-    // 
     async runCommand(exec: string, args: string[], cwd: string = null) {
         this.terminate();
 
@@ -99,69 +101,15 @@ export class UefiTerminal implements vscode.Disposable {
     }
 
     /**
-     * Run a command as a temporary task and wait for it to complete
-     * 
-     * @param exec The process to execute within the task
-     * @param args Arguments to pass to the process. 
-     * @param cwd  The current working directory for the process
-     * @returns    The process exit code
-     */
-    async runCommandAsTask(exec: string, args: string[], name: string = null, cwd: string = null) : Promise<number> {
-        if (!name) { 
-            name = path.basename(exec);
-        }
-
-        // Create a temporary task
-        const task_type = 'uefi-temptask';
-        let task_exec = new vscode.ProcessExecution(exec, args, { cwd: cwd });
-        let task = new vscode.Task({ type: task_type }, name, 'UEFI', task_exec);
-
-        // Create a temporary task provider
-        let taskProvider = vscode.tasks.registerTaskProvider(task_type, {
-            provideTasks: () => {
-                return [task];
-            },
-            resolveTask(_task: vscode.Task) { return undefined; }
-        });
-
-        // Create a promise that completes when the task has ended
-        let promise = new Promise<number>((resolve, reject) => {
-            let evtHandler = vscode.tasks.onDidEndTaskProcess((e) => {
-                if (e.execution.task == task) {
-                    evtHandler.dispose();
-                    resolve(e.exitCode);
-                }
-            });
-        });
-
-        // Execute
-        try {
-            await vscode.tasks.executeTask(task);
-        } finally {
-            // Unregister the task provider so the user can't invoke it
-            taskProvider.dispose();
-        }
-
-        // Wait for task completion
-        return promise;
-    }
-
-    /**
-     * Run python.exe (using the currently configured interpreter)
+     * Run python.exe using the currently configured interpreter (python.pythonPath)
      * 
      * @param args Arguments to pass to python.exe
      * @param cwd  The current working directory
      */
-    // Run a python command in the UEFI terminal window,
-    // using the currently configured python interpreter (python.pythonPath)
     async runPythonCommand(args: string[], label: string = null, cwd: string = null) {
         let pythonPath = utils.getPythonPath(this.workspace);
-
-        let exitCode = await this.runCommandAsTask(pythonPath, args, label, cwd);
-        if (exitCode != 0) {
-            logger.error(`Failed to execute: python ${args} (exit:${exitCode})`);
-            throw Error('Failed to execute python command');
-        }
+        //await this.runShell(pythonPath, args, cwd);
+        await this.runCommand(pythonPath, args, cwd);
     }
 
 }
